requirements.txt
Jinja2==2.6
Markdown==2.2.1
PyYAML==3.10
clint==0.3.1
docopt==0.6.1
nose==1.3.0